Dropped Rest.create(AsyncRestTemplate)
